requirements.txt
mido==1.2.10
pypubsub==4.0.3
python-rtmidi==1.4.9